dbus: request a name on startup
When the service thread is ready, ask a name to dbus
The advantage of doing that is that client applications
can check dbus for bluez-alsa presence, without having
to implement any kinf of polling logic. Also, this
way, they can be notified on the service exit, to
perform any needed cleanup.
Since there must be an instance of bluealsa daemon
per hci device, the dbus name will be
'org.bluez-alsa.hci0', 'org.bluez-alsa.hci1' ...
and so on.

Signed-off-by: Thierry Bultel <thierry.bultel@iot.bzh>